Restructured the file paths and added headers
