+++ .gitignore
@@ -17,3 +17,5 @@ examples/simple_nostorage_scengen/
 
 tests/simulator_tests/test_cases/deterministic_with_network_simulation_output/
 tests/simulator_tests/test_cases/deterministic_simulation_output/
+our_scripts/run_scripts/.DS_Store
+.DS_Store
